Extract iOS app lifecycle support
