+++ apps/ios/Flashcards/Flashcards/App/FlashcardsApp.swift
@@ -1,88 +1,6 @@
 import SwiftUI
 import UIKit
 
-private let flashcardsUITestLaunchScenarioEnvironmentKey: String = "FLASHCARDS_UI_TEST_LAUNCH_SCENARIO"
-private let flashcardsUITestSelectedTabEnvironmentKey: String = "FLASHCARDS_UI_TEST_SELECTED_TAB"
-private let flashcardsUITestAppNotificationTapTypeEnvironmentKey: String = "FLASHCARDS_UI_TEST_APP_NOTIFICATION_TAP_TYPE"
-private let flashcardsUITestAIHandoffCardEnvironmentKey: String = "FLASHCARDS_UI_TEST_AI_HANDOFF_CARD"
-@MainActor
-private var hasConsumedFlashcardsUITestAppNotificationTapEnvironment: Bool = false
-
-@MainActor
-private func runAppBackgroundTask(
-    name: String,
-    operation: @escaping @MainActor () async -> Void
-) {
-    var backgroundTaskIdentifier: UIBackgroundTaskIdentifier = .invalid
-    var operationTask: Task<Void, Never>?
-    var hasBackgroundTaskExpired: Bool = false
-
-    func endBackgroundTaskIfNeeded() {
-        let identifier = backgroundTaskIdentifier
-        backgroundTaskIdentifier = .invalid
-        guard identifier != .invalid else {
-            return
-        }
-
-        UIApplication.shared.endBackgroundTask(identifier)
-    }
-
-    backgroundTaskIdentifier = UIApplication.shared.beginBackgroundTask(withName: name) {
-        hasBackgroundTaskExpired = true
-        operationTask?.cancel()
-        endBackgroundTaskIfNeeded()
-    }
-
-    operationTask = Task { @MainActor in
-        defer {
-            endBackgroundTaskIfNeeded()
-        }
-        guard hasBackgroundTaskExpired == false else {
-            return
-        }
-
-        await operation()
-    }
-}
-
-private enum FlashcardsUITestSelectedTab: String {
-    case review
-    case progress
-    case ai
-    case cards
-    case settings
-
-    var appTab: AppTab {
-        switch self {
-        case .review:
-            return .review
-        case .progress:
-            return .progress
-        case .ai:
-            return .ai
-        case .cards:
-            return .cards
-        case .settings:
-            return .settings
-        }
-    }
-}
-
-private enum FlashcardsUITestAIHandoffCard: String {
-    case firstCard = "first_card"
-}
-
-private enum FlashcardsUITestLaunchError: LocalizedError {
-    case missingAIHandoffCard(FlashcardsUITestAIHandoffCard)
-
-    var errorDescription: String? {
-        switch self {
-        case .missingAIHandoffCard(.firstCard):
-            return "UI test AI handoff requested the first prepared card, but no UI test card is available."
-        }
-    }
-}
-
 private struct CloudSyncPollingTaskID: Hashable {
     let isStartupReady: Bool
     let isSceneActive: Bool
@@ -107,280 +25,6 @@ private func nextProgressContextRolloverDate(now: Date) -> Date {
     }
 
     return nextDay
-}
-
-func appForegroundOperationTabName(tab: AppTab) -> String {
-    switch tab {
-    case .review:
-        return "review"
-    case .progress:
-        return "progress"
-    case .ai:
-        return "ai"
-    case .cards:
-        return "cards"
-    case .settings:
-        return "settings"
-    }
-}
-
-private func appLifecycleTabName(tab: AppTab) -> String {
-    appForegroundOperationTabName(tab: tab)
-}
-
-private func appLifecycleScenePhaseName(scenePhase: ScenePhase) -> String {
-    switch scenePhase {
-    case .active:
-        return "active"
-    case .inactive:
-        return "inactive"
-    case .background:
-        return "background"
-    @unknown default:
-        return "unknown"
-    }
-}
-
-func appForegroundOperationSelectionName(selectedTab: AppTab, previousTab: AppTab?) -> String {
-    let selectedTabName = appForegroundOperationTabName(tab: selectedTab)
-    guard let previousTab,
-          previousTab != selectedTab else {
-        return selectedTabName
-    }
-
-    return "\(appForegroundOperationTabName(tab: previousTab))->\(selectedTabName)"
-}
-
-func appForegroundOperationDurationMilliseconds(startedAt: Date, finishedAt: Date) -> Int {
-    max(0, Int((finishedAt.timeIntervalSince(startedAt) * 1_000).rounded()))
-}
-
-@MainActor
-private func makeAppLifecycleScope(store: FlashcardsStore?) -> IOSObservationScope {
-    let configurationMode: CloudServiceConfigurationMode?
-    if let store {
-        configurationMode = try? store.currentCloudServiceConfiguration().mode
-    } else {
-        configurationMode = nil
-    }
-
-    return IOSObservationScope(
-        feature: .appStartup,
-        userId: store?.cloudSettings?.linkedUserId,
-        workspaceId: store?.workspace?.workspaceId,
-        requestId: nil,
-        clientRequestId: nil,
-        sessionId: nil,
-        runId: nil,
-        cloudState: store?.cloudSettings?.cloudState,
-        configurationMode: configurationMode
-    )
-}
-
-@MainActor
-func makeAppForegroundOperationScope(store: FlashcardsStore) -> IOSObservationScope {
-    IOSObservationScope(
-        feature: .appStartup,
-        userId: store.cloudSettings?.linkedUserId,
-        workspaceId: store.workspace?.workspaceId,
-        requestId: nil,
-        clientRequestId: nil,
-        sessionId: nil,
-        runId: nil,
-        cloudState: store.cloudSettings?.cloudState,
-        configurationMode: try? store.currentCloudServiceConfiguration().mode
-    )
-}
-
-@MainActor
-func logAppForegroundOperationBreadcrumb(
-    action: ForegroundOperationAction,
-    phase: ForegroundOperationPhase,
-    store: FlashcardsStore,
-    selectedTab: AppTab,
-    previousTab: AppTab?,
-    scenePhase: ScenePhase?,
-    isStartupReady: Bool?,
-    isRecoveryGateActive: Bool?,
-    startedAt: Date?,
-    finishedAt: Date?,
-    error: Error?
-) {
-    let scenePhaseName: String?
-    if let scenePhase {
-        scenePhaseName = appLifecycleScenePhaseName(scenePhase: scenePhase)
-    } else {
-        scenePhaseName = nil
-    }
-
-    let durationMilliseconds: Int?
-    if let startedAt,
-       let finishedAt {
-        durationMilliseconds = appForegroundOperationDurationMilliseconds(startedAt: startedAt, finishedAt: finishedAt)
-    } else {
-        durationMilliseconds = nil
-    }
-
-    FlashcardsObservability.addBreadcrumb(
-        .foregroundOperation(
-            ForegroundOperationObservation(
-                scope: makeAppForegroundOperationScope(store: store),
-                action: action,
-                phase: phase,
-                durationMilliseconds: durationMilliseconds,
-                selectedTab: appForegroundOperationSelectionName(selectedTab: selectedTab, previousTab: previousTab),
-                scenePhase: scenePhaseName,
-                isStartupReady: isStartupReady,
-                isRecoveryGateActive: isRecoveryGateActive,
-                cardCount: store.cards.count,
-                deckCount: store.decks.count,
-                pendingOutboxOperationCount: nil,
-                reviewQueueCount: store.reviewQueue.count,
-                reviewDueCount: store.homeSnapshot.dueCount,
-                cloudSyncBlocked: store.isCloudSyncBlocked,
-                errorSummary: error.map { operationError in Flashcards.errorMessage(error: operationError) }
-            )
-        )
-    )
-}
-
-@MainActor
-func prepareVisibleTabForPresentationWithBreadcrumb(
-    store: FlashcardsStore,
-    selectedTab: AppTab,
-    previousTab: AppTab?,
-    scenePhase: ScenePhase?,
-    isStartupReady: Bool?,
-    isRecoveryGateActive: Bool?,
-    now: Date
-) {
-    logAppForegroundOperationBreadcrumb(
-        action: .visibleTabPresentation,
-        phase: .start,
-        store: store,
-        selectedTab: selectedTab,
-        previousTab: previousTab,
-        scenePhase: scenePhase,
-        isStartupReady: isStartupReady,
-        isRecoveryGateActive: isRecoveryGateActive,
-        startedAt: nil,
-        finishedAt: nil,
-        error: nil
-    )
-    store.prepareVisibleTabForPresentation(tab: selectedTab, now: now)
-    logAppForegroundOperationBreadcrumb(
-        action: .visibleTabPresentation,
-        phase: .success,
-        store: store,
-        selectedTab: selectedTab,
-        previousTab: previousTab,
-        scenePhase: scenePhase,
-        isStartupReady: isStartupReady,
-        isRecoveryGateActive: isRecoveryGateActive,
-        startedAt: now,
-        finishedAt: Date(),
-        error: nil
-    )
-}
-
-@MainActor
-private func logAppLifecycleBreadcrumb(
-    action: AppLifecycleAction,
-    store: FlashcardsStore?,
-    stage: String?,
-    scenePhase: ScenePhase?,
-    selectedTab: AppTab?,
-    isStartupReady: Bool?,
-    isRecoveryGateActive: Bool?,
-    messageSummary: String?
-) {
-    let scenePhaseName: String?
-    if let scenePhase {
-        scenePhaseName = appLifecycleScenePhaseName(scenePhase: scenePhase)
-    } else {
-        scenePhaseName = nil
-    }
-
-    let selectedTabName: String?
-    if let selectedTab {
-        selectedTabName = appLifecycleTabName(tab: selectedTab)
-    } else {
-        selectedTabName = nil
-    }
-
-    FlashcardsObservability.addBreadcrumb(
-        .appLifecycle(
-            AppLifecycleObservation(
-                action: action,
-                scope: makeAppLifecycleScope(store: store),
-                stage: stage,
-                scenePhase: scenePhaseName,
-                selectedTab: selectedTabName,
-                isStartupReady: isStartupReady,
-                isRecoveryGateActive: isRecoveryGateActive,
-                messageSummary: messageSummary
-            )
-        )
-    )
-}
-
-@MainActor
-private func consumeFlashcardsUITestAppNotificationTapRequest(
-    processInfo: ProcessInfo,
-    workspaceId: String?
-) -> AppNotificationTapRequest? {
-    guard hasConsumedFlashcardsUITestAppNotificationTapEnvironment == false else {
-        return nil
-    }
-    guard let appNotificationTapType = processInfo.environment[flashcardsUITestAppNotificationTapTypeEnvironmentKey] else {
-        return nil
-    }
-
-    hasConsumedFlashcardsUITestAppNotificationTapEnvironment = true
-    let userInfo: [AnyHashable: Any] = [
-        appNotificationTapTypeUserInfoKey: appNotificationTapType
-    ]
-    let requestIdentifier: String?
-    if appNotificationTapType == AppNotificationTapType.reviewReminder.rawValue,
-       let workspaceId {
-        requestIdentifier = makeReviewNotificationRequestIdentifier(
-            workspaceId: workspaceId,
-            kind: "ui-test",
-            suffix: "launch"
-        )
-    } else {
-        requestIdentifier = nil
-    }
-    return parseAppNotificationTapRequest(
-        userInfo: userInfo,
-        requestIdentifier: requestIdentifier
-    )
-}
-
-@MainActor
-private func makeFlashcardsUITestAIHandoffCard(
-    processInfo: ProcessInfo
-) -> FlashcardsUITestAIHandoffCard? {
-    guard let rawValue = processInfo.environment[flashcardsUITestAIHandoffCardEnvironmentKey] else {
-        return nil
-    }
-
-    return FlashcardsUITestAIHandoffCard(rawValue: rawValue)
-}
-
-@MainActor
-private func makeFlashcardsUITestAIChatPresentationRequest(
-    handoffCard: FlashcardsUITestAIHandoffCard,
-    store: FlashcardsStore
-) throws -> AIChatPresentationRequest {
-    switch handoffCard {
-    case .firstCard:
-        guard let card = store.cards.first else {
-            throw FlashcardsUITestLaunchError.missingAIHandoffCard(.firstCard)
-        }
-
-        return .attachCard(makeAIChatCardReference(card: card))
-    }
 }
 
 @main
